backend/api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
redis
rq
pydantic
requests
httpx
websockets
